requirements.txt
telethon
rich
python-dotenv
aiofiles
streamlit
nest_asyncio
